Rename JS file
