+++ js/index.esm.js
@@ -19,7 +19,7 @@ export { default as Toast } from '../node_modules/bootstrap/js/src/toast.js'
 export { default as Tooltip } from '../node_modules/bootstrap/js/src/tooltip.js'
 export { default as fixModalAriaHidden } from './src/modal.js'
 export { default as photoGalleryGridSlideToImage } from './src/photogallery.js'
-export { default as enableAzNavbarHoverDropdowns } from './src/navbar.js'
+export { default as enableAzNavbarHoverDropdowns } from './src/navbar-hover-dropdown.js'
 
 /**
  * Temporary fix for blocked aria-hidden attribute on modals.
